test: consolidate SSR-user write e2e into a single fixture
The 'caching' and 'optin' fixtures both exercised the same feature — controlling
the SSR user write — but lived apart because they needed opposite global
defaults. Merge them into one 'ssr-user' fixture: the on-by-default routes
(/private, /cacheable, /opted-out) sit alongside an '/optin/**' subtree that is
off-by-default with '/optin/dashboard' opted back in. Using a subtree instead of
a literal '/**' lets both directions coexist in one routeRules block.

The two test files become one 'ssr-user.test.ts' with two describe blocks
('on by default' and 'off by default with per-route opt-in'). No coverage lost:
all cache-guard, opt-out, CVE case-variant, and opt-in cases are retained.
