--- packages/auth0-nuxt/test/fixtures/caching/package.json
@@ -1,5 +1,0 @@
-{
-  "private": true,
-  "name": "caching",
-  "type": "module"
-}
