--- packages/auth0-nuxt/test/fixtures/optin/app.vue
@@ -1,1 +1,0 @@
-<template><NuxtPage /></template>
