--- packages/auth0-nuxt/test/fixtures/optin/pages/opted-in.vue
@@ -1,5 +1,0 @@
-<script setup lang="ts">
-import { useUser } from '../../../../src/runtime/composables/use-user';
-const user = useUser();
-</script>
-<template><div><span data-testid="user-sub">{{ user?.sub ?? 'anonymous' }}</span></div></template>
