+++ packages/auth0-nuxt/test/fixtures/ssr-user/app.vue
