+++ packages/auth0-nuxt/test/fixtures/ssr-user/package.json
@@ -1,5 +1,5 @@
 {
   "private": true,
-  "name": "optin",
+  "name": "ssr-user",
   "type": "module"
 }
