+++ packages/auth0-nuxt/test/fixtures/ssr-user/pages/opted-out.vue
