+++ packages/auth0-nuxt/test/fixtures/ssr-user/pages/private.vue
